# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEXTAN Costing & Quote Generator API",
    "description": "Internal costing sheets and client-facing branded quote exports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/organisations": {
      "get": {
        "tags": [
          "Organisations"
        ],
        "summary": "List Organisations",
        "description": "Purpose: List all active (non-deleted) organisations; supports optional name search.\nInputs: q (Optional[str]) \u2014 name substring filter\nOutputs: list[OrganisationRead]\nOwner: [Claude]",
        "operationId": "list_organisations_api_v1_organisations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Name search filter",
              "title": "Q"
            },
            "description": "Name search filter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationRead"
                  },
                  "title": "Response List Organisations Api V1 Organisations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organisations"
        ],
        "summary": "Create Organisation",
        "description": "Purpose: Create a new organisation in the shared client database.\nInputs: OrganisationCreate (name, address)\nOutputs: OrganisationRead\nOwner: [Claude]",
        "operationId": "create_organisation_api_v1_organisations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/autocomplete": {
      "get": {
        "tags": [
          "Organisations"
        ],
        "summary": "Autocomplete Organisations",
        "description": "Purpose: Typeahead autocomplete \u2014 returns id+name only for lightweight dropdown.\nInputs: q (str) \u2014 search prefix\nOutputs: list[OrganisationAutocomplete]\nOwner: [Claude]",
        "operationId": "autocomplete_organisations_api_v1_organisations_autocomplete_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationAutocomplete"
                  },
                  "title": "Response Autocomplete Organisations Api V1 Organisations Autocomplete Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{org_id}": {
      "get": {
        "tags": [
          "Organisations"
        ],
        "summary": "Get Organisation",
        "description": "Purpose: Retrieve a single organisation by ID (active only).\nInputs: org_id (str UUID path param)\nOutputs: OrganisationRead\nOwner: [Claude]",
        "operationId": "get_organisation_api_v1_organisations__org_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organisations"
        ],
        "summary": "Update Organisation",
        "description": "Purpose: Update name or address of an organisation.\nInputs: org_id (str UUID), OrganisationUpdate (name, address \u2014 both optional)\nOutputs: OrganisationRead\nOwner: [Claude]",
        "operationId": "update_organisation_api_v1_organisations__org_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organisations"
        ],
        "summary": "Delete Organisation",
        "description": "Purpose: Soft-delete an organisation (set deleted_at). FK on CostingSheet remains intact.\n         Hard delete is blocked to protect CostingSheet relational history.\nInputs: org_id (str UUID)\nOutputs: 204 No Content\nOwner: [Claude]",
        "operationId": "delete_organisation_api_v1_organisations__org_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganisationAutocomplete": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "OrganisationAutocomplete",
        "description": "Purpose: Lightweight item for autocomplete dropdown.\nInputs: N/A (response body)\nOutputs: id, name\nOwner: [Claude]"
      },
      "OrganisationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganisationCreate",
        "description": "Purpose: Body for POST /organisations.\nInputs: name (str), address (Optional[str])\nOutputs: N/A\nOwner: [Claude]"
      },
      "OrganisationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "created_by_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Created By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_by_user_id",
          "created_at",
          "updated_at"
        ],
        "title": "OrganisationRead",
        "description": "Purpose: Organisation details in list and get responses.\nInputs: N/A (response body)\nOutputs: id, name, address, created_by_user_id, created_at, updated_at\nOwner: [Claude]"
      },
      "OrganisationUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "title": "OrganisationUpdate",
        "description": "Purpose: Body for PUT /organisations/{org_id}.\nInputs: name (Optional[str]), address (Optional[str])\nOutputs: N/A\nOwner: [Claude]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}